+++ bcdi/preprocessing/process_scan.py
@@ -5,7 +5,10 @@
 #   (c) 07/2019-05/2021 : DESY PHOTON SCIENCE
 #       authors:
 #         Jerome Carnis, jerome.carnis@esrf.fr
-"""Workflow for BCDI data preprocessing of a single scan, before phase retrieval."""
+"""Workflow for BCDI data preprocessing of a single scan, before phase retrieval.
+
+The detector is expected to be on a goniometer.
+"""
 
 import gc